More extensive threading support
A lot of apps still don’t use proper thread IDs, but most do use titles. This change checks for a valid title and then uses that as a "thread ID" to lump similiar notifications together. With this change, Facebook Messenger and Hangouts now support stacking based on the sender’s name.